third_party/blink/web_tests/platform/generic/http/tests/security/isolatedWorld/sandboxed-iframe-expected.txt

CONSOLE ERROR: Blocked script execution in 'http://127.0.0.1:8000/security/isolatedWorld/resources/fail.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: PASS
Test that executing scripts in an isolated world works even inside sandboxed iframes. The test passes, if an alert with the text "PASS" is shown.
